using modal for gpus and training sweep to attempt to reproduce results

+++ src/evals/metrics/utils.py
@@ -95,8 +95,9 @@ def evaluate_probability(model, batch):
     avg_losses = losses / num_token_gt
     normalized_probs = torch.exp(-avg_losses)
 
-    avg_losses = avg_losses.cpu().numpy().tolist()
-    normalized_probs = normalized_probs.cpu().numpy().tolist()
+    # NumPy does not support bfloat16 tensors directly.
+    avg_losses = avg_losses.to(dtype=torch.float32).cpu().numpy().tolist()
+    normalized_probs = normalized_probs.to(dtype=torch.float32).cpu().numpy().tolist()
     return [
         {"prob": prob, "avg_loss": avg_loss}
         for prob, avg_loss in zip(normalized_probs, avg_losses)
